+++ src/parse-changelog.js
@@ -28,15 +28,15 @@ const todoItems = (content) => {
 export const parseChangelog = (changelogContent) => {
   const hasContent = !!changelogContent.trim();
   if (!hasContent) {
-    return { version: -1, items: [] };
+    return { version: '-1', items: [] };
   }
   const content = '\n' + changelogContent;
   const newLineAndNewVersionString = '\n' + LINE_START_FOR_NEW_VERSION;
   if (!content.includes(newLineAndNewVersionString)) {
-    return { version: -1, items: [] };
+    return { version: '-1', items: [] };
   }
   const versions = content.split(newLineAndNewVersionString);
   const firstVersionParagraph = versions[1];
-  const version = Number.parseInt(firstVersionParagraph.split('\n')[0], 10);
+  const version = firstVersionParagraph.split('\n')[0];
   return { version, items: todoItems(firstVersionParagraph) };
 };